Store more data in a human-readable fashion
This is optional and is disabled by default.

More item data (in addition to inventories) can be serialized.  Doing so
yields the risk of future incompatibilities.  It may come in handy where
human-readability is desirable.  If parts of the item cannot be
serialized (because this patch is not intended to be exhaustive), the
remaining data will be stored in the usual base64 format.